Bug fixes; consent form
User consent to data collection is more explicit.
Bug where last report suite determined the number of eVars/props/events
in the slot overview has been fixed.

Reportsuites chosen not to be exported are listed in the Excel as well.
The number of sheets in the Excel formula is output.