Massive File size optimizations
I found some low-hanging fruit to dramatically lower the file size of js
files used. The biggest impact is in using a modified wsse which doesn't
import all 300KB of crypto.

The essentials.js file is much more important now, as it holds all
modules. Browserify shim is used to help this process. This will allow
for more code to be cached.

Module Licenses are exported to a seperate file.